Build: Establish CMake targets, tests, and CI boundaries
- Split shared build configuration into focused warnings, options, and sanitizer targets
- Convert production modules to linkable static targets with explicit dependencies
- Remove the global source include path and enforce architecture boundaries
- Make tests link production targets and inherit strict compiler warnings
- Add plugin-contract CI coverage and install the plugin SDK artifacts

+++ test/component/net_transport_test.cpp
@@ -351,7 +351,7 @@ TEST_F(TcpStreamTest, ReadSome_ReturnsAvailableBytes) {
     auto n = stream.read_some(buf);
     ASSERT_TRUE(n);
     EXPECT_EQ(*n, 5);
-    EXPECT_EQ(str(std::vector<std::uint8_t>(buf.begin(), buf.begin() + *n)), "hello");
+    EXPECT_EQ(str(std::vector<std::uint8_t>(buf.begin(), buf.begin() + static_cast<std::ptrdiff_t>(*n))), "hello");
 }
 
 TEST_F(TcpStreamTest, ConnectionRefused_ReturnsConnectionFailed) {
